backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-spa
whitenoise
pillow
uritemplate
pyyaml
six
requests